Rename methods.
- shell.SplitQ -> texts.SplitLikeShellString
- shell.QuotedFirstWord -> texts.FirstWord
